requirements.txt
langchain
langchain-core
openai
google-generativeai
python-dotenv
pandas
jupyter
langchain_community
langchain-openai
matplotlib